[REF] vscode: refactor odooPath processing
goal :
odooPath processing must be accessible in utils,
It must be generic
