+++ vscode/client/common/utils.ts
@@ -1,4 +1,9 @@
-import { ExtensionContext, Uri, Webview } from "vscode";
+import { ExtensionContext, Uri, Webview, workspace} from "vscode";
+import * as fs from 'fs';
+import { URI } from "vscode-languageclient";
+import untildify from 'untildify';
+import * as readline from 'readline';
+
 
 /**
  * A helper function which will get the webview URI of a given file or resource.
@@ -30,4 +35,63 @@ export async function getCurrentConfig(context: ExtensionContext) {
     const configs: any = context.globalState.get("Odoo.configurations");
     const activeConfig: number = Number(context.workspaceState.get('Odoo.selectedConfiguration'));
     return (configs && activeConfig > -1 ? configs[activeConfig] : null);
+}
+
+export async function evaluateOdooPath(odooPath){
+	const workspaceFolders = workspace.workspaceFolders;
+
+	const fillTemplate = (template, vars = {}) => {
+		const handler = new Function('vars', [
+			'const tagged = ( ' + Object.keys(vars).join(', ') + ' ) =>',
+			'`' + template + '`',
+			'return tagged(...Object.values(vars))'
+		].join('\n'));
+		const res = handler(vars)
+		return res;
+	};
+
+
+	for (const i in workspaceFolders){
+		const folder = workspaceFolders[i];
+		let PATH_VAR_LOCAL = global.PATH_VARIABLES;
+		// TODO we should reevaluate path variables each time
+		PATH_VAR_LOCAL["workspaceFolder"] = folder.uri.path;
+		odooPath = fillTemplate(odooPath,PATH_VAR_LOCAL);
+		const version = await getOdooVersion(odooPath);
+		if (version){
+			global.PATH_VARIABLES["workspaceFolder"] = folder.uri.path; 
+			return {"path": odooPath,"version": version};
+		}
+	}
+	return null;
+}
+
+export async function getOdooVersion(odooPath: URI) {
+	let versionString = null;
+	const releasePath = untildify(odooPath) + '/odoo/release.py';
+	if (fs.existsSync(releasePath)) {
+		const rl = readline.createInterface({
+			input: fs.createReadStream(releasePath),
+			crlfDelay: Infinity,
+		});
+
+		for await (const line of rl) {
+			if (line.startsWith('version_info')) {
+				versionString = line;
+				// Folder is invalid if we don't find any version info
+				if (!versionString) {
+					return null;
+				} else {
+					// Folder is valid if a version was found
+					const versionRegEx = /\(([^)]+)\)/; // Regex to obtain the info in the parentheses
+					const versionArray = versionRegEx.exec(versionString)[1].split(', ');
+					const version = `${versionArray[0]}.${versionArray[1]}.${versionArray[2]}` + (versionArray[3] == 'FINAL' ? '' : ` ${versionArray[3]}${versionArray[4]}`);
+					return version;
+				}
+			}
+		}
+	} else {
+		// Folder is invalid if odoo/release.py was never found
+		return null;
+	}
 }
